Receive callbacks from the sensor list if new values of the currently selected sensor arrive or if the selected sensor changes.
Each time one of this events happen the gui (graph and values area) will be updated
Receive value update callback from active sensors while measuring

If the list receives new values from the currently selected sensor it forwards the information to the main window via another callback. This way the graph view and the values area of the gui can be updated.
If the currently selected sensor changes the graph will completely be redrawn with the data of the new sensor.
Start/Stop value worker and receive its values

The values are stored in an internal values object that automatically calculates max, min, avg as well as it manages a history of values.
After receiving new values the parent sensor list is being notified
Create a demoValueWorker that sends random values and the corresponding timestamp via queue to its parent
Refactor timeWorker to use a queue instead of callbacks